// forms/src/lib/form-navigation/types.ts
export interface RoutableFieldOptions {
    /**
     * The route that represents the field within the form,
     * this will be autogenerated based on key / index but
     * can be overridden in configuration if custom urls are
     * needed
     *
     * @type {string}
     * @memberof RoutableFieldOptions
     */
    route?: string;
}

export interface RouteableField {
    /**
     * The route to this portion of the form
     *
     * @type {string}
     * @memberof RouteableField
     */
    formRoute?: string;

    /**
     * The direct route to a given field
     *
     * @type {string}
     * @memberof RouteableField
     */
    fieldRoute?: string;
}